Refactor pretty printing machinery to separate files.
